\name{GPSmle}
\alias{GPSmle}
\title{
\code{\link{GPSmle.default}}
}
\description{
\code{\link{GPSmle.default}}
}
\usage{
GPSmle(data, group = rep(1:2, each = 5), 
type = c("pvalue", "normalization", "ecdf"), 
method = c("GP-Theta", "Lowess", "GP-Quantile", "Quantile", "TMM"), 
maxIter = 500, paired = FALSE, ncpu = 1, geneid = NULL, empirical.T.stats = NULL)
}
\arguments{
  \item{data}{
\code{\link{GPSmle.default}}
}
  \item{group}{
\code{\link{GPSmle.default}}
}
  \item{type}{
\code{\link{GPSmle.default}}
}
  \item{method}{
\code{\link{GPSmle.default}}
}
  \item{maxIter}{
\code{\link{GPSmle.default}}
}
  \item{paired}{
\code{\link{GPSmle.default}}
}
  \item{ncpu}{
\code{\link{GPSmle.default}}
}
  \item{geneid}{
\code{\link{GPSmle.default}}
}
  \item{empirical.T.stats}{
\code{\link{GPSmle.default}}}
}
\details{
\code{\link{GPSmle.default}}
}
\seealso{
\code{\link{GPSmle.default}}
}

