\name{summary.GPSmle}
\alias{summary.GPSmle}
\title{
Summary of GPSmle
}
\description{
Summary of GPSmle
}
\usage{
\method{summary}{GPSmle}(object, ...)
}
\arguments{
  \item{object}{
the GPSmle object returned by \code{\link{GPSmle}} or \code{\link{deGPS_mRNA}}.
}
  \item{\dots}{
see the \code{\link{summary.default}}
}
}
\details{
summary of the GPSmle
}
\value{
summary of the GPSmle
}
